etl/requirements.txt
fastf1>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
meteostat>=1.6.6
pyyaml>=6.0.1
